changing uint8_t to int16_t to prevent weird double ball, back n forth issue.

+++ headers/main/brick.h
@@ -1,6 +1,6 @@
 #ifndef BRICK_HEADER 
 #define BRICK_HEADER set
-uint8_t CheckTopOrBottomBrick(uint8_t x, uint8_t  y);
+uint8_t CheckTopOrBottomBrick(int16_t x, int16_t  y);
 void UpdateBrick(uint8_t side, uint8_t x, uint8_t  y);
 void DrawLevelBricks(uint8_t *level);
